fix(dashboard): rewrite FILTER aggregates for PostgreSQL compat
Replace aggregate FILTER clauses with CASE WHEN inside aggregates in
the findings aging and actions aging queries (Query 3b and Query 4) so
GET /api/dashboard/summary works on PostgreSQL versions that do not
support the FILTER syntax. Also add ::int cast to the periodDays
parameter in customerApiKeys usage query to resolve the date >= integer
operator error.

+++ src/api/routes/customerApiKeys.ts
@@ -192,7 +192,7 @@ router.get(
            FROM api_keys k
            LEFT JOIN api_usage_daily u
              ON u.api_key_id = k.id
-             AND u.date >= CURRENT_DATE - $2
+             AND u.date >= CURRENT_DATE - $2::int
            WHERE k.organization_id = $1
            GROUP BY k.id, k.label, k.status
            ORDER BY total_requests DESC`,
